Replacing Monitoring Service with configurable MonitoringServiceFactory
- Created Monitoring Service interface
- Created factory to allow for different types of monitoring services via configuration
- Created LocalMonitoringService that does not try to publish to cloudwatch
- Renamed the old Monitoring Service to AwsCloudWatchMonitoringServiceImpl
