+++ transitclock/src/main/java/org/transitclock/monitoring/MonitoringServiceFactory.java
@@ -1,0 +1,44 @@
+/*
+ * This file is part of Transitime.org
+ *
+ * Transitime.org is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License (GPL) as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * any later version.
+ *
+ * Transitime.org is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with Transitime.org .  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+package org.transitclock.monitoring;
+
+import org.transitclock.config.StringConfigValue;
+import org.transitclock.utils.ClassInstantiator;
+
+public class MonitoringServiceFactory {
+
+    // The name of the class to instantiate
+    private static StringConfigValue className =
+            new StringConfigValue("transitclock.monitoring.MonitoringServiceClass",
+                    "org.transitclock.monitoring.AwsCloudWatchMonitoringServiceImpl",
+                    "Specifies the name of the class used for generating " +
+                            "Monitoring data.");
+
+    private static MonitoringService singleton = null;
+
+    /********************** Member Functions **************************/
+
+    public static MonitoringService getInstance() {
+        // If the MonitoringService hasn't been created yet then do so now
+        if (singleton == null) {
+            singleton = ClassInstantiator.instantiate(className.getValue(), MonitoringService.class);
+        }
+
+        return singleton;
+    }
+}
